# Lab2/pzpi-22-3-shpak-denys-lab2/docs/swagger.yaml
basePath: /
definitions:
  dtos.CompanyDTO:
    properties:
      address:
        description: |-
          Address is the address of the company.
          Example: "Koningin Wilhelminaplein 13, 1062 HH Amsterdam"
        type: string
      creator:
        allOf:
        - $ref: '#/definitions/dtos.UserDTO'
        description: Creator is the user who created the company.
      deliveries:
        description: Deliveries is the list of deliveries that belong to the company.
        items:
          $ref: '#/definitions/dtos.DeliveryDTO'
        type: array
      id:
        description: |-
          ID is the unique identifier of the company.
          Example: 1
        type: integer
      name:
        description: |-
          Name is the name of the company.
          Example: "TechTeam"
        type: string
      routes:
        description: Routes is the list of routes that belong to the company.
        items:
          $ref: '#/definitions/dtos.RouteDTO'
        type: array
      users:
        description: Users is the list of users that belong to the company.
        items:
          $ref: '#/definitions/dtos.UserDTO'
        type: array
    type: object
  dtos.DeliveryDTO:
    properties:
      company_id:
        description: |-
          CompanyID is the unique identifier of the company
          Example: 1
        type: integer
      date:
        description: |-
          Date is the date of the delivery
          Example: 2021-01-01T00:00:00Z
        type: string
      duration:
        description: |-
          Duration is the duration of the delivery
          Example: 1h
        type: string
      id:
        description: |-
          ID is the unique identifier of the delivery
          Example: 1
        type: integer
      products:
        description: Products is the list of products of the delivery
        items:
          $ref: '#/definitions/dtos.ProductDTO'
        type: array
      route_id:
        description: |-
          RouteID is the unique identifier of the route
          Example: 1
        type: integer
      status:
        description: |-
          Status is the status of the delivery
          Example: completed
        type: string
    type: object
  dtos.ProductCategoryDTO:
    properties:
      description:
        description: |-
          Description is the product category description
          Example: Food products
        type: string
      id:
        description: |-
          ID is the product category identifier
          Example: 1
        type: integer
      is_perishable:
        description: |-
          IsPerishable is a flag that indicates if the product category is perishable
          Example: true
        type: boolean
      max_humidity:
        description: |-
          MaxHumidity is the maximum humidity that the product category can be stored
          Example: 20
        type: number
      max_temperature:
        description: |-
          MaxTemperature is the maximum temperature that the product category can be stored
          Example: 20
        type: number
      min_humidity:
        description: |-
          MinHumidity is the minimum humidity that the product category can be stored
          Example: 10
        type: number
      min_temperature:
        description: |-
          MinTemperature is the minimum temperature that the product category can be stored
          Example: 10
        type: number
      name:
        description: |-
          Name is the product category name
          Example: Food
        type: string
    type: object
  dtos.ProductDTO:
    properties:
      id:
        description: |-
          ID is the unique identifier of the product
          Example: 1
        type: integer
      name:
        description: |-
          Name is the name of the product
          Example: Beer
        type: string
      product_category:
        allOf:
        - $ref: '#/definitions/dtos.ProductCategoryDTO'
        description: ProductCategory is the category of the product
      weight:
        description: |-
          Description is the description of the product
          Example: A cold beer
        type: number
    type: object
  dtos.RouteDTO:
    properties:
      details:
        description: |-
          Details is the details of the route
          Example: Route 1 details
        type: string
      id:
        description: |-
          ID is the unique identifier of the route
          Example: 1
        type: integer
      name:
        description: |-
          Name is the name of the route
          Example: Route 1
        type: string
      status:
        description: |-
          Status is the status of the route
          Example: normal temperature
        type: string
      waypoints:
        description: Waypoints is a list of waypoints that the route has
        items:
          $ref: '#/definitions/dtos.WaypointDTO'
        type: array
    type: object
  dtos.SensorDataDTO:
    properties:
      date:
        description: |-
          Date is the date of the SensorData
          Example: 2021-01-01T00:00:00Z
        type: string
      humidity:
        description: |-
          Humidity is the humidity of the SensorData
          Example: 0.5
        type: number
      id:
        description: |-
          ID is the unique identifier of the SensorData
          Example: 1
        type: integer
      mean_pressure:
        description: |-
          WindDirection is the wind direction of the SensorData
          Example: 0.5
        type: number
      temperature:
        description: |-
          Temperature is the temperature of the SensorData
          Example: 25.5
        type: number
      wind_speed:
        description: |-
          Rain is the rain of the SensorData
          Example: 0.5
        type: number
    type: object
  dtos.UserDTO:
    properties:
      id:
        description: |-
          ID is the user identifier
          Example: 1
        type: integer
      name:
        description: |-
          Name is the user name
          Example: John Doe
        type: string
      password:
        description: |-
          Email is the user email
          Example: password123
        type: string
    type: object
  dtos.WaypointDTO:
    properties:
      details:
        description: |-
          Details is the details of the Waypoint
          Example: Waypoint requires diagnostics
        type: string
      device_serial:
        description: |-
          DeviceSerial is the serial number of the device that sent the Waypoint
          Example: 123456
        type: string
      get_weather_alerts:
        description: |-
          GetWeatherAlerts is a boolean that indicates if the company should get weather alerts
          Example: true
        type: boolean
      id:
        description: |-
          ID is the unique identifier of the Waypoint
          Example: 1
        type: integer
      latitude:
        description: |-
          Latitude is the latitude of the Waypoint
          Example: -12.045
        type: number
      longitude:
        description: |-
          Longitude is the longitude of the Waypoint
          Example: -77.0311
        type: number
      name:
        description: |-
          Name is the name of the Waypoint
          Example: Waypoint 1
        type: string
      send_data_frequency:
        description: |-
          SendDataFrequency is the frequency of the data sent by the device in minutes
          Example: 5
        type: integer
      sensor_data:
        description: Altitude is the altitude of the Waypoint
        items:
          $ref: '#/definitions/dtos.SensorDataDTO'
        type: array
      status:
        description: |-
          Status is the status of the Waypoint
          Example: ok
        type: string
    type: object
  handlers.AddUserToCompanyRequest:
    properties:
      role:
        allOf:
        - $ref: '#/definitions/handlers.Role'
        description: Role is the role of the user in the company
        example: user | admin | manager
      userID:
        description: |-
          UserID is the ID of the user to add
          Example: 1
        type: integer
    type: object
...
  handlers.CompanyRequest:
    properties:
      address:
        description: |-
          Address is the address of the company
          Example: 123 Main St
        type: string
      name:
        description: |-
          Name is the name of the company
          Example: Wayra
        type: string
    type: object
...
  handlers.RemoveUserFromCompanyRequest:
    properties:
      userID:
        description: |-
          UserID is the ID of the user to remove
          Example: 1
        type: integer
    type: object
...
  handlers.Role:
    enum:
    - user
    - admin
    - manager
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAdmin
    - RoleManager
...
  handlers.UpdateUserInCompanyRequest:
    properties:
      role:
        description: Role is the role of the user in the company
        example: user | admin | manager
        type: string
      userID:
        description: |-
          UserID is the ID of the user to update
          Example: 1
        type: integer
    type: object
  handlers.UpdateUserRequest:
    properties:
      name:
        description: |-
          Name is the name of the user
          Example: John Doe
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  termsOfService: http://swagger.io/terms/
  title: API Specification
  version: "1.0"
paths:
...
  /company:
    post:
      consumes:
      - application/json
      description: Registers a new company with the provided details
      parameters:
      - description: Company details
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/handlers.CompanyRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Register a new company
      tags:
      - company
  /company/:
    get:
      description: Retrieves a list of all companies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.CompanyDTO'
            type: array
      security:
      - BearerAuth: []
      summary: Get all companies
      tags:
      - company
  /company/{company_id}:
    delete:
      description: Deletes a company by its ID
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete a company
      tags:
      - company
    get:
      description: Retrieves the details of a company by its ID
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get company details
      tags:
      - company
    put:
      consumes:
      - application/json
      description: Updates the details of an existing company
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: integer
      - description: Updated company details
        in: body
        name: company
        required: true
        schema:
          $ref: '#/definitions/handlers.CompanyRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update company details
      tags:
      - company
  /company/{company_id}/add-user:
    post:
      consumes:
      - application/json
      description: Adds a user to a company if the request is made by the company
        creator
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: integer
      - description: User ID to add
        in: body
        name: userID
        required: true
        schema:
          $ref: '#/definitions/handlers.AddUserToCompanyRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Add a user to a company
      tags:
      - company
  /company/{company_id}/remove-user:
    delete:
      consumes:
      - application/json
      description: Removes a user from a company if the request is made by the company
        creator
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: integer
      - description: User ID to remove
        in: body
        name: userID
        required: true
        schema:
          $ref: '#/definitions/handlers.RemoveUserFromCompanyRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Remove a user from a company
      tags:
      - company
  /company/{company_id}/update-user:
    put:
      consumes:
      - application/json
      description: Updates a user in a company if the request is made by the company
        creator
      parameters:
      - description: Company ID
        in: path
        name: company_id
        required: true
        type: integer
      - description: User ID to update
        in: body
        name: userID
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserInCompanyRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update a user in a company
      tags:
      - company
...
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      operationId: delete-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Get a user by ID
      operationId: get-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Get a user
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update a user by ID
      operationId: update-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUserRequest'
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Update a user
      tags:
      - user
...
securityDefinitions:
  BearerAuth:
    description: 'Provide the Bearer token in the format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"